+++ CHANGELOG.md
@@ -7,9 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
 ### Changed
 
 - Improve error message when it is not possible to access the data storage
+- Remove `chunk_size` parameter from configuration
+- Try to take Intake URL from an environment var first  
 
 ## [1.5.0] - 2023-09-19